Update README.md
Updated README.md file by adding necessary information about missing algorithms and ordered the table according to the folders. Also edited a typo in the 6th entry of table "EfficientNet" which was earlier "EffecientNet". Replaced the Table Header "Misc" by "Description" as I find it more relevant according to the information of the field. The algorithms which are missing from the previous version are as follows:-
1. Deepface Ensemble
2. CNN
3. PCA
4. SIFT